refactor: make get_block_status common for methods
And remove duplicate BlockStatus.
