+++ crates/pathfinder/src/rpc/v02/method/get_transaction_receipt.rs
@@ -1,9 +1,9 @@
 use anyhow::Context;
 
-use crate::core::{StarknetBlockNumber, StarknetTransactionHash};
-use crate::rpc::v02::types::reply::BlockStatus;
+use crate::core::StarknetTransactionHash;
+use crate::rpc::v02::common::get_block_status;
 use crate::rpc::v02::RpcContext;
-use crate::storage::{RefsTable, StarknetBlocksTable, StarknetTransactionsTable};
+use crate::storage::{StarknetBlocksTable, StarknetTransactionsTable};
 
 #[derive(serde::Deserialize, Debug, PartialEq, Eq)]
 pub struct GetTransactionReceiptInput {
@@ -80,22 +80,6 @@ pub async fn get_transaction_receipt(
     });
 
     jh.await.context("Database read panic or shutting down")?
-}
-
-/// Determines block status based on the current L1-L2 stored in the DB.
-fn get_block_status(
-    db_tx: &rusqlite::Transaction<'_>,
-    block_number: StarknetBlockNumber,
-) -> anyhow::Result<BlockStatus> {
-    // All our data is L2 accepted, check our L1-L2 head to see if this block has been accepted on L1.
-    let l1_l2_head =
-        RefsTable::get_l1_l2_head(db_tx).context("Read latest L1 head from database")?;
-    let block_status = match l1_l2_head {
-        Some(number) if number >= block_number => BlockStatus::AcceptedOnL1,
-        _ => BlockStatus::AcceptedOnL2,
-    };
-
-    Ok(block_status)
 }
 
 mod types {
@@ -532,7 +516,8 @@ mod types {
 mod tests {
     use super::*;
     use crate::core::{
-        ContractAddress, EventData, EventKey, Fee, StarknetBlockHash, StarknetTransactionHash,
+        ContractAddress, EventData, EventKey, Fee, StarknetBlockHash, StarknetBlockNumber,
+        StarknetTransactionHash,
     };
     use crate::{starkhash, starkhash_bytes};
